feat/creating a counter of habits

+++ app/api/habits/[id]/route.ts
@@ -1,9 +1,34 @@
 import { type NextRequest, NextResponse } from "next/server"
 import { prisma } from "@/lib/prisma"
+import {
+  UpdateHabitSchema,
+  type UpdateHabitSchemaType
+} from "@/lib/schema/habit"
+import { auth } from "@clerk/nextjs/server"
 
 export async function DELETE(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await params
+
+    const { userId } = await auth()
+
+    if (!userId) {
+      return NextResponse.json({
+        error: "Unauthorized"
+      }, { status: 401 })
+    }
+
+    const userDb = await prisma.user.findFirst({
+      where: {
+        clerkUserId: userId
+      }
+    })
+
+    if (!userDb) {
+      return NextResponse.json({
+        error: "user not find on db"
+      }, { status: 401 })
+    }
 
     await prisma.habit.delete({
       where: {
@@ -22,33 +47,95 @@ export async function DELETE(request: NextRequest, { params }: { params: Promise
 export async function PATCH(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await params
+    
+    const { userId } = await auth()
+
+    if (!userId) {
+      return NextResponse.json({
+        error: "Unauthorized"
+      }, { status: 401 })
+    }
+
+    const userDb = await prisma.user.findFirst({
+      where: {
+        clerkUserId: userId
+      }
+    })
+
+    if (!userDb) {
+      return NextResponse.json({
+        error: "user not find on db"
+      }, { status: 401 })
+    }
 
     const body = await request.json()
+
+    const parsedBody = UpdateHabitSchema.safeParse(body)
+
+    if (!parsedBody.success) throw new Error(parsedBody.error.message)
+    
     const {
       name,
       emoji,
-      goal,
-      motivation,
-      reminder,
+      endDate,
+      startDate,
+      reminder, 
       frequency,
       color,
-      startDate,
-      endDate
-    } = body
+      clock,
+      limitCounter,
+      status,
+      goal
+    } = parsedBody.data
+
+    const existingGoal = await prisma.goals.findFirst({
+      where: {
+        id: goal
+      }
+    })
+
+    if(existingGoal) {
+      await prisma.habit.update({
+        where: {
+          id
+        },
+        data: {
+          goals: {
+            disconnect: {
+              id: existingGoal.id
+            }
+          }
+        }
+      })
+    }
+
+    //QUANDO ATUALIZAR UM HÁBITO, EU DEVO PERCORRER MEUS COMPLETIONS DO HABITO E ATUALIZAR O LIMIT COUNTER
 
     const newHabit = await prisma.habit.update({
       where: {
         id
       },
       data: {
-        name: name,
-        emoji: emoji,
-        reminder: reminder,
-        frequency: frequency,
-        color: color,
-        startDate: startDate,
-        endDate: endDate
-      }
+        userId: userDb.id,
+        name,
+        emoji,
+        startDate: new Date(startDate),
+        endDate: endDate ? new Date(endDate) : null,
+        reminder,
+        frequency, // Json
+        color,
+        limitCounter,
+        updatedAt: new Date(),
+        ...(goal  && {goals: {
+          connect: {
+            id: goal
+          }
+        }}),
+        clock
+      },
+      include: {
+        completions: true,
+      },
     })
 
     console.log(newHabit, "updating")
